backend/requirements.txt
uvicorn==0.25.0
fastapi==0.106.0
pandas==2.1.4
langchain==0.0.353
tiktoken==0.5.2
pydantic==2.5.3
faiss-cpu==1.7.4
langchain-community==0.1.0
google-api-python-client
